More tests and fixed bugs on PageMenu
