+++ android/fastlane/Fastfile
@@ -26,7 +26,7 @@ platform :android do
     #sh('echo $PWD || false')
     sh("find ../app/build/outputs/apk/ -name \\*.apk -exec rm {} \\;")
     android_set_version_code(
-      version_code: "109"
+      version_code: "110"
     )
     android_set_version_name(
       version_name: "0.0.11"
